### 2023-05-23 #73 @AugustinChrtn
[Re] Exploration in Model-based Reinforcement Learning by Empirically Estimating Learning Progress 
**Original article:** Lopes, M., Lang, T., Toussaint, M., & Oudeyer, P. Y. (2012). Exploration in model-based reinforcement learning by empirically estimating learning progress. Advances in neural information processing systems, 25.

**Scientific domain:** Computational Neuroscience
