### 2023-06-07 #73 @oliviaguest
👋 Hi @hhihn @kkhetarpal @ghost-nn-machine @MA-Ramirez can any/all of you review this for ReScience C? 😊  Please let me know as soon as possible if you can take this on, I would appreciate that of course, but also be aware I might be harder to reach than usual till next week. Thank you!
