### 2023-08-04 #73 @MehdiKhamassi
Dear editors, Dear @rougier, Dear @oliviaguest,
Any news from the reviewing process?
I wish you all a great summer!
Best regards
mehdi
