### 2023-08-29 #73 @kkhetarpal
Dear ReScience/submissions

Thank you for the reviewer invitation. Unfortunately, I am unable to commit
to this at this point due to several other constraints.
Sorry about that. Thanks for understanding.

Kind Regards
Khimya

On Mon, Aug 28, 2023 at 1:06 AM Alejandra Ramirez ***@***.***>
wrote:

> Dear all,
>
> thank you very much for the invitation to review the article.
> Nonetheless, this time I won’t be able to review it due to time
> constraints.
>
> All the best
>
