### 2023-09-17 #73 @HaoZeke
I would be happy to review this. However, we should have a "compute requirements" section as well in the submission process to better understand what resources reviewers might need.
