### 2024-03-18 #73 @mo-arvan
In this submission, the authors attempt to reproduce the work of Lopes et al (Exploration in Model-based Reinforcement Learning by Empirically Estimating Learning Progress). They discuss the challenges, discrepancies and potential solutions for resolving the identified issues. While the authors of the original paper have not released their source code, they have assisted the authors of this paper and have clarified some aspect of the work. This paper does not fit into my area of expertise, I have attempted to read, understand, and follow the instructions of the authors.

I started with forking and cloning the source code released by submission 73. List of dependencies used by authors are rather short, which reduces the chance of discrepancy management issues (dependency hell). Containerizing the environment to run everything inside a docker container was straightforward. The `dockerfile`, the steps I took to reproduce the results, and the output of the scripts ran are added to the [fork](https://github.com/mo-arvan/rescience-submission-73) inside a directory called `artifact`.

The repository released by authors is sufficiently documented and properly structured. It took approximately two hours to run all the scripts on a modern workstation. While re-running the scripts provided by authors was easy, sifting through the results to find the corresponding figure was more convoluted than it should be. To put it simply, the code generates too many figures and naming schema is not clear enough to identify the matching figure in the paper. This is not a major issue, but they authors could refactor the naming schema to include the figure number, if a figure is included in their paper.

The authors extensively discuss the their findings and the challenges they have faced. In particular, Table 2 provides a great overview of all the challenges, the resolutions, and the results. Although they did not manage to successfully reproduce all, their investigation is thorough and the clarifications provided by the original authors are helpful. Overall, I believe this submission complements the original paper.

Question for the authors:

Can you provide some information on how to match the figures generated by the code to the figures in the paper?
