### 2024-04-08 14f6dcc
Modified the structure of some directories. Added a text to describe what each program to launch does and an estimate of the computation time.
