+++ main_functions.py
@@ -4,7 +4,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 import time
 import itertools
-from multiprocessing import Pool
+from multiprocessing import Pool, cpu_count
 
 from lopesworld import Lopes_environment
 
@@ -12,6 +12,7 @@ from policy_functions import value_iteration, get_optimal_policy
 from policy_functions import policy_evaluation, get_agent_current_policy
 from agents import Rmax, BEB, EpsilonMB, EBLP, RmaxLP
 
+plt.switch_backend("Agg")
 
 def loading_environments():
     """Load all the environment previously generated by the 'generation_env.py' file."""
@@ -161,7 +162,8 @@ def main_function(all_seeds, every_simulation, play_params, agent_parameters):
     else:
         all_parameters = [[play_params, all_seeds[index_seed], agent_parameters[index_seed],
                            every_simulation[index_seed]] for index_seed in range(len(all_seeds))]
-    pool = Pool()
+    nb_processes_to_use = max(1,cpu_count()-1)
+    pool = Pool(processes=nb_processes_to_use)
     results = pool.map(one_parameter_play_function, all_parameters)
     pool.close()
     pool.join()
@@ -171,7 +173,7 @@ def main_function(all_seeds, every_simulation, play_params, agent_parameters):
         rewards[result[0]] = result[1][0]
         real_pol_errors[result[0]] = result[1][2]
     time_after = time.time()
-    print('Computation time: '+str(time_after - before))
+    print('Computation time: '+str(round(time_after - before))+" seconds.")
     return opt_pol_errors, real_pol_errors, rewards
 
 
@@ -247,21 +249,26 @@ def save_and_plot(optimal_stats, real_stats, rewards_stats, agents_tested,
 
     x_axis = [i*play_parameters["step_between_VI"] for i in range(length_pol)]
     plot_agents(agents_tested, pol_opti, SEM_pol_opti, x_axis, colors, markers, marker_sizes)
-    plt.savefig('Results/pol_error_opti'+time_end +
+    plt.title("Policy errors with the optimal policy on the model each agent learned")
+    plt.savefig('Plots/pol_error_opti'+time_end +
                 names_environments[0]+'.pdf', bbox_inches='tight')
+    plt.close()
 
     plot1D([-13, 0.5], "Steps", "Policy value error")
     plot_agents(agents_tested, pol_real, SEM_pol_real, x_axis, colors, markers, marker_sizes)
-    plt.savefig('Results/pol_error_real'+time_end +
+    plt.title("Policy error with the different agent policies")
+    plt.savefig('Plots/pol_error_real'+time_end +
                 names_environments[0]+'.pdf', bbox_inches='tight')
     plt.close()
 
     x_reward = [i for i in range(play_parameters["trials"])]
     plot1D([-1, 26], "Trials", "Reward")
     plot_agents(agents_tested, reward, SEM_reward, x_reward, colors, markers, marker_sizes)
-    plt.savefig('Results/Rewards'+time_end+names_environments[0]+'.pdf', bbox_inches='tight')
-    print('Saved the new plots in the Results folder')
+    plt.title("Rewards for each agent")
+    plt.savefig('Plots/Rewards'+time_end+names_environments[0]+'.pdf', bbox_inches='tight')
     plt.close()
+    
+    print('Saved the new plots in the Plots folder')
 
 
 def plot1D(ylim, xlabel, ylabel):
