+++ parameter_fitting.py
@@ -9,6 +9,7 @@ import copy
 
 from matplotlib.colors import TwoSlopeNorm
 
+plt.switch_backend("Agg")
 
 def range_parameters_agent(list1, list2):
     list_of_params = []
@@ -185,181 +186,198 @@ def fit_parameters_agent(environments, name_agent, nb_iters, first_hp, second_hp
 
 
 # Parameter fitting for each agent
+if __name__ == "__main__":
+    print("""This program launches all the parameter fitting.
+          
+          There are 19 parameter fits with different agent or parameter ranges.
+          
+          For each experiment, we print the computational time each fit took.
 
-environments_parameters = loading_environments()
-play_params = {'trials': 100, 'max_step': 30, 'screen': False, 'photos': [
-    10, 20, 50, 80, 99], 'accuracy_VI': 0.01, 'step_between_VI': 50}
+          This code uses multiprocessing and runs faster on machines with many
+          cores. 
 
+          An estimation of the total computational time is 30 times the time
+          of the first experiment (e.g 100 min if the first experiment takes 200
+          seconds.)
 
-# Reproduction of Lopes et al. (2012)
+          Parameter fitting can take a few hours to run.
 
-environments = ['Lopes']
-nb_iters = 20
-
-agent_parameters = {
-    'ε-greedy': {'gamma': 0.95, 'epsilon': 0.3},
-    'R-max': {'gamma': 0.95, 'm': 8, 'Rmax': 1, 'm_u': 12, 'condition': 'informative'},
-    'BEB': {'gamma': 0.95, 'beta': 7, 'coeff_prior': 2, 'condition': 'informative'},
-    'ζ-EB': {'gamma': 0.95, 'beta': 3, 'step_update': 10, 'alpha': 0.3, 'prior_LP': 0.01},
-    'ζ-R-max': {'gamma': 0.95, 'Rmax': 1, 'm': 2, 'step_update': 10, 'alpha': 0.3, 'prior_LP': 0.01}}
-
-
-agent_name = 'R-max'
-starting_seed = 10000
-
-first_hp = ['m']+[1]+[i for i in range(5, 41, 5)]
-second_hp = ['m_u']+[1]+[i for i in range(5, 41, 5)]
-fit_parameters_agent(environments, agent_name, nb_iters, first_hp, second_hp,
-                     {agent_name: agent_parameters[agent_name]}, starting_seed, play_params)
+          """)
+    
+    environments_parameters = loading_environments()
+    play_params = {'trials': 100, 'max_step': 30, 'screen': False, 'photos': [
+        10, 20, 50, 80, 99], 'accuracy_VI': 0.01, 'step_between_VI': 50}
 
 
-first_hp = ['m']+[i for i in range(2, 15, 2)]
-second_hp = ['m_u']+[i for i in range(2, 21, 2)]
-fit_parameters_agent(environments, agent_name, nb_iters, first_hp, second_hp,
-                     {agent_name: agent_parameters[agent_name]}, starting_seed, play_params)
+    # Reproduction of Lopes et al. (2012)
+
+    environments = ['Lopes']
+    nb_iters = 20
+
+    agent_parameters = {
+        'ε-greedy': {'gamma': 0.95, 'epsilon': 0.3},
+        'R-max': {'gamma': 0.95, 'm': 8, 'Rmax': 1, 'm_u': 12, 'condition': 'informative'},
[…]
