+++ Optimal-policies/README.md
@@ -1,0 +1,1 @@
+This directory contains the theoretical values and optimal policies for each cell of all the environments used.
