+++ Plots/README.md
@@ -1,0 +1,1 @@
+We store the plots generated by the main.py in this folder. The plots we show in the article are stored in the Article subfolder.
