+++ requirements.yml
@@ -1,0 +1,8 @@
+name: re_chartounyetal
+channels:
+  - defaults
+dependencies:
+  - python=3.9.13
+  - matplotlib=3.5.2
+  - numpy=1.21.5
+  - seaborn=0.11.2

[…]
