### 2024-04-09 6667179
Modified the structure of the repository and the places were the different plots or data were stored.
