+++ parameter_fitting.py
@@ -11,11 +11,19 @@ from matplotlib.colors import TwoSlopeNorm
 
 plt.switch_backend("Agg")
 
+name_to_save = {'ε-greedy': 'e-greedy',
+                'BEB': 'BEB',
+                'R-max': 'R-max',
+                'ζ-EB': 'zeta-EB',
+                'ζ-R-max': 'zeta-R-max'}
+
+
 def range_parameters_agent(list1, list2):
     list_of_params = []
     for elem1 in range(1, len(list1)):
         for elem2 in range(1, len(list2)):
-            list_of_params.append({list1[0]: list1[elem1], list2[0]: list2[elem2]})
+            list_of_params.append(
+                {list1[0]: list1[elem1], list2[0]: list2[elem2]})
     return list_of_params
 
 
@@ -60,7 +68,8 @@ def get_best_performance(pol_error, name_agent, first_param, second_param, range
     array_result = np.zeros((len(first_param)-1, len(second_param)-1))
     for index_hp_1, hp_1 in enumerate(first_param[1:]):
         for index_hp_2, hp_2 in enumerate(second_param[1:]):
-            array_result[(index_hp_1, index_hp_2)] = avg_pol_error[name_agent, hp_1, hp_2]
+            array_result[(index_hp_1, index_hp_2)
+                         ] = avg_pol_error[name_agent, hp_1, hp_2]
     return array_result
 
 
@@ -69,28 +78,27 @@ def save_results_parameter_fitting(pol_opti, SEM_pol_opti, pol_real, SEM_pol_rea
                                    play_parameters, name_environments):
     """Save all the results of the parameter fitting in a '.npy' format."""
     time_end = str(round(time.time() % 1e7))
-    np.save('Parameter fitting/Data/'+name_agent+'_' +
-            name_environments[0]+'_pol_opti_'+time_end+'.npy', pol_opti)
-    np.save('Parameter fitting/Data/'+name_agent+'_' +
-            name_environments[0]+'_SEM_pol_opti_'+time_end+'.npy', SEM_pol_opti)
-    np.save('Parameter fitting/Data/'+name_agent+'_' +
-            name_environments[0]+'_pol_real_'+time_end+'.npy', pol_real)
-    np.save('Parameter fitting/Data/'+name_agent+'_' +
-            name_environments[0]+'_SEM_pol_real_'+time_end+'.npy', SEM_pol_real)
+    tmp_path = 'Parameter fitting/Data/'+name_to_save[name_agent]+'/'
+    path = tmp_path+name_agent+'_' + name_environments[0]
+    np.save(path+'_pol_opti_'+time_end+'.npy', pol_opti)
+    np.save(path+'_SEM_pol_opti_'+time_end+'.npy', SEM_pol_opti)
+    np.save(path+'_pol_real_'+time_end+'.npy', pol_real)
+    np.save(path+'_SEM_pol_real_'+time_end+'.npy', SEM_pol_real)
     return time_end
 
 
 def plot_from_saved(name_agent, first_param, second_param,
                     step_between_VI, name_environments, time_end, optimal):
+    path = 'Parameter fitting/Data/'+name_to_save[name_agent]+'/'
     if optimal:
-        pol = np.load('Parameter fitting/Data/'+name_agent+'_' + name_environments[0] +
+        pol = np.load(path+name_agent+'_' + name_environments[0] +
                       '_pol_opti_'+time_end+'.npy', allow_pickle=True)[()]
-        SEM_pol = np.load('Parameter fitting/Data/'+name_agent+'_' + name_environments[0] +
+        SEM_pol = np.load(path+name_agent+'_' + name_environments[0] +
                           '_SEM_pol_opti_'+time_end+'.npy', allow_pickle=True)[()]
     else:
[…]
