+++ main_functions.py
@@ -17,34 +17,44 @@ plt.switch_backend("Agg")
 def loading_environments():
     """Load all the environment previously generated by the 'generation_env.py' file."""
     env_param = {}
-    malus_0_1 = np.load('Environments/Rewards_-0.1.npy', allow_pickle=True)
-    transitions_0_1 = np.load('Environments/Transitions_-0.1_1.npy', allow_pickle=True)
-    env_param['Lopes'] = {'transitions': transitions_0_1, 'rewards': malus_0_1}
+    path_rewards_0_1 = 'Environments/Rewards_-0.1.npy'
+    malus_0_1 = np.load(path_rewards_0_1, allow_pickle=True)
+    path_transitions_0_1 = 'Environments/Transitions_-0.1_1.npy'
+    transitions_0_1 = np.load(path_transitions_0_1, allow_pickle=True)
+    env_param['Lopes'] = {'transitions': transitions_0_1,
+                          'rewards': malus_0_1}
 
     for index_non_stat in range(1, 21):
 
         transitions_non_stat_article = np.load(
-            'Environments/Transitions_non_stat_article_-0.1_1_' + str(index_non_stat) + '.npy',
+            'Environments/Transitions_non_stat_article_-0.1_1_' +
+            str(index_non_stat) + '.npy',
             allow_pickle=True)
 
         transitions_strong_non_stat = np.load(
-            'Environments/Transitions_strong_non_stat_-0.1_1_'+str(index_non_stat)+'.npy',
+            'Environments/Transitions_strong_non_stat_-0.1_1_' +
+            str(index_non_stat)+'.npy',
             allow_pickle=True)
 
         env_param["Lopes_non_stat_article_{0}".format(index_non_stat)] = {
-            'transitions': transitions_0_1, 'rewards': malus_0_1,
+            'transitions': transitions_0_1,
+            'rewards': malus_0_1,
             'transitions_after_change': transitions_non_stat_article}
 
         env_param["Lopes_strong_non_stat_{0}".format(index_non_stat)] = {
-            'transitions': transitions_0_1, 'rewards': malus_0_1,
+            'transitions': transitions_0_1,
+            'rewards': malus_0_1,
             'transitions_after_change': transitions_strong_non_stat}
 
     malus_1 = np.load('Environments/Rewards_-1.npy', allow_pickle=True)
+
     for nb_env in range(1, 11):
-        transitions_1 = np.load('Environments/Transitions_-1_' + str(nb_env) + '.npy',
-                                allow_pickle=True)
+        transitions_1 = np.load(
+            'Environments/Transitions_-1_' + str(nb_env) + '.npy',
+            allow_pickle=True)
         env_param['Stationary_-1_' + str(nb_env)] = {
-            'transitions': transitions_1, 'rewards': malus_1}
+            'transitions': transitions_1,
+            'rewards': malus_1}
         for index_non_stat in range(1, 11):
 
             transitions_non_stat_article = np.load(
@@ -58,10 +68,12 @@ def loading_environments():
                 allow_pickle=True)
 
             env_param["Non_stat_article_-1_{0}".format(nb_env)+'_{0}'.format(index_non_stat)] = {
-                'transitions': transitions_0_1, 'rewards': malus_1,
+                'transitions': transitions_0_1,
+                'rewards': malus_1,
                 'transitions_after_change': transitions_non_stat_article}
             env_param["Strong_non_stat_-1_{0}".format(nb_env)+'_{0}'.format(index_non_stat)] = {
-                'transitions': transitions_0_1, 'rewards': malus_1,
+                'transitions': transitions_0_1,
+                'rewards': malus_1,
                 'transitions_after_change': transitions_strong_non_stat}
 
     return env_param
[…]
