+++ Parameter fitting/1DPlots/README.md
@@ -1,1 +1,1 @@
-This directory contains the 1D parameter fitting plots.
+This directory contains the 1D parameter fitting plots. The plots are saved in the folder with the agent name. The plots of our experiments are in the Article folder.
