+++ Plots/OptimalPolicy/README.md
@@ -1,0 +1,1 @@
+Contains the policy value differences between the optimal policy on the model learned by the agent and the theoretical optimal policy on the real model of the environment.
