+++ ValueIterationPolicies/README.md
@@ -1,0 +1,1 @@
+We save in this directory the values of each cell depending on the policy the agent chose. In the OptimalPolicies folder, we show the theoretical best policies for all the environments generated.
