+++ Parameter fitting/1DPlots/zeta-R-max/README.md
@@ -1,0 +1,1 @@
+1D plots for the parameter fitting of zeta-R-max.
