### 2024-04-09 eba924f
Matched the plot names with the figures from the article.
