+++ parameter_fitting.py
@@ -88,7 +88,7 @@ def save_results_parameter_fitting(pol_opti, SEM_pol_opti, pol_real, SEM_pol_rea
 
 
 def plot_from_saved(name_agent, first_param, second_param,
-                    step_between_VI, name_environments, time_end, optimal):
+                    step_between_VI, name_environments, time_end, optimal,title=''):
     path = 'Parameter fitting/Data/'+name_to_save[name_agent]+'/'
     if optimal:
         pol = np.load(path+name_agent+'_' + name_environments[0] +
@@ -101,11 +101,11 @@ def plot_from_saved(name_agent, first_param, second_param,
         SEM_pol = np.load(path+name_agent+'_' + name_environments[0] +
                           '_SEM_pol_real_'+time_end+'.npy', allow_pickle=True)[()]
     plot_parameter_fitting(pol, SEM_pol, name_agent, first_param,
-                           second_param, step_between_VI, name_environments, time_end, optimal)
+                           second_param, step_between_VI, name_environments, time_end, optimal,title)
 
 
 def plot_parameter_fitting(pol, SEM_pol, name_agent, first_param, second_param,
-                           step_between_VI, name_environments, time_end, optimal):
+                           step_between_VI, name_environments, time_end, optimal,title=''):
     """Plot 1D curves for parameter fitting."""
     markers = ['^', 'o', 'x', '*', 's']
     colors = ['#9d02d7', '#0000ff', "#ff7763", "#ffac1e", "#009435"]
@@ -114,13 +114,15 @@ def plot_parameter_fitting(pol, SEM_pol, name_agent, first_param, second_param,
 
     plot_2D(array_avg_pol_last_500, first_param, second_param)
     path = 'Parameter fitting/Heatmaps/'+name_to_save[name_agent]+'/'
+    if title!='':
+        path = path + title
     if optimal:
         plt.title(name_agent+' optimal policy - last 500 steps')
-        plt.savefig(path+name_agent+' optimal_policy ' +
+        plt.savefig(path+name_agent+'_optimal_policy ' +
                     name_environments[0]+time_end+'.pdf', bbox_inches='tight')
     else:
         plt.title(name_agent+' agent policy - last 500 steps')
-        plt.savefig(path+name_agent+' real_policy ' +
+        plt.savefig(path+name_agent+'_real_policy_' +
                     name_environments[0]+time_end+'.pdf', bbox_inches='tight')
     plt.close()
 
@@ -144,16 +146,18 @@ def plot_parameter_fitting(pol, SEM_pol, name_agent, first_param, second_param,
                 plt.legend()
                 path = 'Parameter fitting/1DPlots/' + \
                     name_to_save[name_agent] + '/'
+                if title!='':
+                    path = path+title
                 if optimal:
                     plt.title(name_agent+" optimal policy with " +
                               str(first_param[0])+" = "+str(hp_1))
-                    plt.savefig(path+'pol_error_opti_' + name_agent + '_' +
+                    plt.savefig(path+'pol_error_opti_' + name_agent+ '_' +
                                 name_environments[0] + str(hp_2)+"_"
                                 + str(time.time())+'.pdf', bbox_inches='tight')
                 else:
                     plt.title(name_agent+" agent policy with " +
                               str(first_param[0])+" = "+str(hp_1))
-                    plt.savefig(path+'pol_error_real_'+name_agent + '_' +
+                    plt.savefig(path+'pol_error_real_'+name_agent+ '_' +
[…]
