+++ Data/README.md
@@ -1,2 +1,6 @@
-This directory contains the main result figures shown in the article.
+This directory contains the main data of the article obtained with main.py. The folder **Article** contains the data we use to generate the plots of the article. 
 
+We saved data for three metrics:
+* The folder **Rewards** contains the rewards the different agent obtained
+* The folder **AgentPolicy** contains the data of the difference between the agent policy and the optimal policy
+* The folder **OptimalPolicy** contains the data of the difference between the best policy computed on the model the agent learned and the optimal policy.
