+++ Plots/README.md
@@ -1,1 +1,6 @@
-We store the plots generated by the main.py in this folder. The plots we show in the article are stored in the Article subfolder.
+We store the plots generated by the main.py in this folder. The folder **Article** contains the plots we show in the article. 
+
+We show plots for three metrics:
+* The folder **Rewards** contains the rewards the different agent obtained
+* The folder **AgentPolicy** contains the plots of the difference between the agent policy and the optimal policy
+* The folder **OptimalPolicy** contains the plots of the difference between the best policy computed on the model the agent learned and the optimal policy.
