+++ Parameter fitting/README.md
@@ -1,1 +1,1 @@
-This folder contains the plots and data for the parameter fitting. The results obtained in the article are saved in the Article folder. The new results are saved in the three other folders.
+This folder contains the plots and data for the parameter fitting. The results obtained in the article are saved in the Article folder. The new results are saved in the three other folders (either 1D plots, not very used in the article, data, or heatmaps in the supplementary section).

[…]
