### 2024-04-09 #73 @AugustinChrtn
Dear @mo-arvan, thank you for your detailled feedback, the docker and your question.

I modified the structure of the repository and clarified how to match each plot to a Figure of the article. I separated the different metrics (or agents for parameter fitting) in different folders. In addition, the plot names the code generates now have the number of the figure they correspond to in their names. I also added a clearer output for the computational time and for the figures that the code is generating online. Thank you for pointing this out and I hope that these changes will make the results of the code easier to reuse and understand. Please let me know if I should change anything else in the code or in the article!

Dear @rougier, I answered to your issue and I hope that you can run main now. Please let me know if you cannot.
