### 2024-04-10 #73 @rougier
I've read the article and I would like to congratulate for this work. The paper is really really clear and explain both the different difficulties, how you manage to solve problems and it represents a fair amount of work. I've got only a small minor problem during testing but it has been fixed since then. I've only some some minor points:

1. Your requirements file is bit strict in terms of version for the different library. On the one hand this ensures reproducibility (which is good) but in the other hand you need to install a specific python version which I found a bit annoying. This does not need correction and you precise specification is ok. Just to let you know.

2. You need to add a license to your code because no license means no right for anyone to do anything with your code (see https://choosealicense.com/). We have no recommendation for ReScience as long as it is an open license.

The "major" point concerns the title/conclusion. The semantic is [Re] for replication, [¬Re] for failed replication and [~Re] for approximate replication which I think correspond to your case since you did not manage to replicate all the results in spite of a lot of efforts you put in this work. The idea with the title is to quickly warn the reader that the work is not totally reproducible. If you disagree with the change [Re] to [~Re], we can discuss it.

One this is done, I think the paper can be published. I'll let @oliviaguest decides on the next step.
