### 2024-04-11 #73 @AugustinChrtn
Dear @rougier, thank you very much for your feedback. We did spend a lot of time and effort to replicate this paper and we're very happy that you liked it. I answer to all of your points below:

Minor points:
1. I agree that the requirement file is rather strict. It was to make sure that people would be able to generate the exact same plots and results as the ones we present.
2. Thank you for pointing it out. I added an MIT license.

Major point: I do agree that the paper is ~Re rather than Re. I changed the metadata (and the title).

We would like to thank again the two reviewers for their feedback and we are ready to proceed depending on @oliviaguest 's instructions.
