### 2024-06-12 c2b4e30
Added information before publication
