+++ metadata.yaml
@@ -5,7 +5,7 @@
 #  - For a successful replication, it should be prefixed with "[Re]"
 #  - For a failed replication, it should be prefixed with "[¬Re]"
 #  - For other article types, no instruction (but please, not too long)
-title: "[Re]Exploration in Model-based Reinforcement Learning by Empirically Estimating Learning Progress"
+title: "[~Re]Exploration in Model-based Reinforcement Learning by Empirically Estimating Learning Progress"
 
 # List of authors with name, orcid number, email and affiliation
 # Affiliation "*" means contact author
