### 2024-08-12 #73 @MehdiKhamassi
Dear Editors, Dear @rougier , Dear @oliviaguest ,
This is my yearly message asking if there is any news from the reviewing process? :-)
I wish you all a great summer!
Best regards
mehdi
