### 2024-09-02 #73 @rougier
Sorry for the loooong delay. I'll publish the paper today. @AugustinChrtn Do you have a link to the latex files, it'll make my life easier.
