### 2024-09-02 #73 @AugustinChrtn
Sure! I downloaded the source files from overleaf. Please let me know if you need something else!
Augustin

[
[Rescience Chartouny.zip](https://github.com/user-attachments/files/16834643/Rescience.Chartouny.zip)
](url)
