### 2024-09-02 #73 @rougier
Thanks. Can you check https://sandbox.zenodo.org/records/105346 if everything looks right? This is a sandbox version, not the final publication.
