### 2024-09-02 #73 @rougier
It's online https://zenodo.org/records/13627804 !!! Congratulations!
